Added functionality of restarting media player and more implementation for the delete audio section

+++ src/controllers/DeleteAudio.java
@@ -2,6 +2,8 @@ package controllers;
 
 import Services.DirectoryServices;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Button;
 import javafx.scene.control.ListView;
 
 import java.io.IOException;
@@ -12,64 +14,87 @@ public class DeleteAudio extends Controller {
 
     private String _previousscene = "/fxml/MainMenu.fxml";
     private String _subdirectory;
+    @FXML private Button _deletebutton;
     @FXML private ListView<String> _audiosubdirs;
     @FXML private ListView<String> _audiofiles;
 
 
 
-    public void SelectAudioSubDirectory()   {
+    public void SelectAudioSubDirectory() {
 
         _audiofiles.getItems().clear();
 
         _subdirectory = _audiosubdirs.getSelectionModel().getSelectedItem();
 
-        //Check for null;
-        ArrayList<String> audiofiles = DirectoryServices.listAudio(_subdirectory);
-        for(String s: audiofiles) {
-            _audiofiles.getItems().add(s);
-        }
 
-     }
+        if (_subdirectory == null) {
+
+
+        } else {
+
+            ArrayList<String> audiofiles = DirectoryServices.listAudio(_subdirectory);
+            for (String s : audiofiles) {
+                _audiofiles.getItems().add(s);
+            }
+
+        }
+    }
 
 
      public void DeleteAudiofile() throws IOException, InterruptedException {
 
-        //Have a pup up for confirmation;
+         //Have a pup up for confirmation;
+
 
          String audio = _audiofiles.getSelectionModel().getSelectedItem();
 
-         String cmd = "rm -f "+ "./audio/"+ _subdirectory+"/" + audio;
-         ProcessBuilder deleteaudio = new ProcessBuilder("/bin/bash", "-c", cmd);
-         Process deleteprocess = deleteaudio.start();
-         deleteprocess.waitFor();
+         if (audio == null) {
 
-         ArrayList<String> audiofiles = DirectoryServices.listAudio(_subdirectory);
-         _audiofiles.getItems().clear();
+             CreateAlert(Alert.AlertType.WARNING, "No item selected", "ERROR You have not selected an item");
 
-         for(String s: audiofiles) {
-             _audiofiles.getItems().add(s);
+
+         } else {
+
+             String cmd = "rm -f " + "./audio/" + _subdirectory + "/" + audio;
+             ProcessBuilder deleteaudio = new ProcessBuilder("/bin/bash", "-c", cmd);
+             Process deleteprocess = deleteaudio.start();
+             deleteprocess.waitFor();
+
+             ArrayList<String> audiofiles = DirectoryServices.listAudio(_subdirectory);
+             _audiofiles.getItems().clear();
+
+             for (String s : audiofiles) {
+                 _audiofiles.getItems().add(s);
+             }
+
+             ProcessBuilder deletedir = new ProcessBuilder("sh", "-c", "./scripts/checkdir_isempty.sh" + " \"" + _subdirectory + "\"");
+             Process deletedirprocess = deletedir.start();
+             deletedirprocess.waitFor();
+
+             _audiosubdirs.getItems().clear();
+
+
+             ArrayList<String> _audiosubfolders = DirectoryServices.listAudioFolders();
+             for (String s : _audiosubfolders) {
+                 _audiosubdirs.getItems().add(s);
+             }
+
          }
-
-         ProcessBuilder deletedir = new ProcessBuilder("sh", "-c", "./scripts/checkdir_isempty.sh"+" "+_subdirectory);
-         Process deletedirprocess = deletedir.start();
-
-         _audiosubdirs.getItems().clear();
-
-
-         //NOT UPDATING PROPERLY;
-         System.out.println("SDGHALGH");
-         ArrayList<String> _audiosubfolders = DirectoryServices.listAudioFolders();
-         for(String s: _audiosubfolders) {
-             _audiosubdirs.getItems().add(s);
-         }
-
      }
 
 
 
      @FXML
-     public void initialize(){
+     public void initialize() throws InterruptedException, IOException {
 
+
+         // Remove all empty directories.
+         ProcessBuilder deletedirs = new ProcessBuilder("sh", "-c", "./scripts/delete_all_nonemptydirs.sh");
+         Process deletedirprocess = deletedirs.start();
+         deletedirprocess.waitFor();
+
+
+        // Populate ListView
          ArrayList<String> _audiosubfolders = DirectoryServices.listAudioFolders();
 
          for(String s: _audiosubfolders) {
@@ -77,8 +102,6 @@ public class DeleteAudio extends Controller {
          }
 
      }
-
-
